# apps/backend/docs/public-api/openapi.yaml
openapi: 3.0.3
info:
  title: Social Media Scheduler Public API
  version: 1.0.0
  description: |
    External-facing REST API for third-party integrations with the Social Media Scheduler platform.
    
    ## Authentication
    
    All requests require an API key passed in the `X-API-Key` header:
    
    ```
    X-API-Key: sk_live_your_api_key_here
    ```
    
    API keys can be created and managed through the dashboard at `/settings/api-keys`.
    
    ## Rate Limiting
    
    API requests are rate-limited per API key. Default limits:
    - 1000 requests per hour per API key
    - 5000 requests per hour per workspace (across all keys)
    
    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed in the window
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets
    
    When rate limited, you'll receive a `429 Too Many Requests` response with a `Retry-After` header.
    
    ## Scopes & Permissions
    
    API keys have granular permission scopes that control access to endpoints:
    
    - `posts:read` - Read posts and drafts
    - `posts:write` - Create, update, and delete posts
    - `analytics:read` - Read analytics data
    - `media:read` - Read media library
    - `media:write` - Upload and delete media
    - `accounts:read` - Read connected social accounts
    - `accounts:write` - Connect and disconnect accounts
...
    
...
    
...
servers:
...
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/public/v1/analytics:
    get:
      tags:
        - Analytics
      summary: Get aggregated analytics
      description: Retrieve aggregated analytics data for your workspace
      operationId: getAnalytics
      security:
        - ApiKeyAuth: [analytics:read]
      parameters:
        - name: startDate
          in: query
          description: Start date for analytics period (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date for analytics period (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: platform
          in: query
          description: Filter by social media platform
          required: false
          schema:
            type: string
            enum: [twitter, facebook, instagram, linkedin, tiktok, google_business]
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregatedAnalytics'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Format: `sk_live_<random_string>`
        
        Create API keys in your dashboard at `/settings/api-keys`.
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
        example: 1000
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
        example: 999
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
        example: 1709812800
  schemas:
...
    AggregatedAnalytics:
      type: object
      properties:
        analytics:
          type: object
          properties:
            totals:
              type: object
              properties:
                impressions:
                  type: integer
                  description: Total impressions across all posts
                engagements:
                  type: integer
                  description: Total engagements (likes + comments + shares + clicks)
                likes:
                  type: integer
                  description: Total likes
                comments:
                  type: integer
                  description: Total comments
                shares:
                  type: integer
                  description: Total shares
                clicks:
                  type: integer
                  description: Total clicks
            period:
              type: object
              properties:
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
            postsAnalyzed:
              type: integer
              description: Number of posts included in analytics
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: Unauthorized
        message:
          type: string
          description: Human-readable error message
          example: Invalid API key
        code:
          type: string
          description: Machine-readable error code
          example: INVALID_API_KEY
        requestId:
          type: string
          description: Unique request identifier for debugging
          example: req_abc123xyz
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        docsUrl:
          type: string
          description: Link to error documentation
          example: https://docs.example.com/errors/INVALID_API_KEY
  responses:
...
    Unauthorized:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid API key
            code: INVALID_API_KEY
            requestId: req_abc123xyz
            timestamp: '2026-03-07T10:30:00Z'
            docsUrl: https://docs.example.com/errors/INVALID_API_KEY
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: 'Missing required scopes: posts:write'
            code: INSUFFICIENT_SCOPE
            requestId: req_abc123xyz
            timestamp: '2026-03-07T10:30:00Z'
            docsUrl: https://docs.example.com/errors/INSUFFICIENT_SCOPE
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
            example: 3600
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Too Many Requests
            message: Rate limit exceeded. Please retry after 3600 seconds.
            code: RATE_LIMIT_EXCEEDED
            requestId: req_abc123xyz
            timestamp: '2026-03-07T10:30:00Z'
            docsUrl: https://docs.example.com/errors/RATE_LIMIT_EXCEEDED